Standardized the return for pod wait to ScenarioResult
Now the return from the wait includes ScenarioResult and an error attribute.

In case of success, the ScenarioResult contains the provider pods and their hit/cold details, otherwise it contains both *some* of the pods that may have hit ready and the pods that never hit unready.

In case of failures, the intermediate results are printed in addition to the logs being dumped.

Signed-off-by: Gloire Rubambiza <gloire@ibm.com>